==> scripts/10 join everything.R <==
#join everything
#aoo <- read.csv("./results/final_with_aooeoo.csv", row.names = 1)
#aoo <- read.csv("./results/aoo_veg.csv", row.names = 1)
aoo <- read.csv("./results/aoo_veg_cites.csv", row.names = 1)
use <- read.csv("./results/Use_results_general.csv", row.names = 1)
ucs <- read.csv("./results/tabla_UCs.csv", row.names = 1)
names(aoo)
names(use)
names(ucs)
use <- use %>% rename(nombre = especies)
final <- left_join(aoo, use) %>% left_join(ucs)
View(final)

write.csv(final, "./results/final_results.csv")
data <- read.csv("./results/final_results.csv", row.names = 1)
data %<>% mutate(aoo2 = 4*aoo2, aoo10 = 100*aoo10)
write.csv(data, "./results/final_results_corrected_AOO.csv")
